+++ app/src/main/java/org/example/Singup.java
@@ -1,0 +1,14 @@
+// package org.example;
+// import javax.swing.JFrame;
+
+// // public class Singup extends JFrame {
+// //     // singUp(){
+// //     //     setSize(850,800);
+// //     //     setVisible(true);
+
+// //     }
+//     public static void main(String[] args) {
+//         // new Singup();
+        
+//     }
+// }
