+++ app/src/main/resources/icons/bit.jpg
